Fix merge boom
