+++ Assets/Prefab/Enemy/InfectorNear/Enemy_Zombie.prefab
@@ -12,7 +12,7 @@ GameObject:
   - component: {fileID: 641467605861483491}
   - component: {fileID: 641467605861483484}
   - component: {fileID: 7825429015952538566}
-  - component: {fileID: 4134463714415941971}
+  - component: {fileID: 2254360005788471185}
   m_Layer: 0
   m_Name: Enemy_Zombie
   m_TagString: UnTouchEnemy
@@ -128,8 +128,8 @@ Rigidbody2D:
   m_SleepingMode: 1
   m_CollisionDetection: 0
   m_Constraints: 0
---- !u!70 &4134463714415941971
-CapsuleCollider2D:
+--- !u!61 &2254360005788471185
+BoxCollider2D:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
@@ -142,5 +142,15 @@ CapsuleCollider2D:
   m_UsedByEffector: 0
   m_UsedByComposite: 0
   m_Offset: {x: 0, y: 0}
+  m_SpriteTilingProperty:
+    border: {x: 0, y: 0, z: 0, w: 0}
+    pivot: {x: 0.5, y: 0.5}
+    oldSize: {x: 0.31, y: 0.48}
+    newSize: {x: 0.31, y: 0.48}
+    adaptiveTilingThreshold: 0.5
+    drawMode: 0
+    adaptiveTiling: 0
+  m_AutoTiling: 0
+  serializedVersion: 2
   m_Size: {x: 0.31, y: 0.48}
-  m_Direction: 0
+  m_EdgeRadius: 0
