+++ Assets/Prefab/Player/Rook/Rook.prefab
@@ -12,7 +12,6 @@ GameObject:
   - component: {fileID: 8437023491213692342}
   - component: {fileID: 8437023491213692341}
   - component: {fileID: 8437023491213692338}
-  - component: {fileID: 1932466399221208136}
   - component: {fileID: 5934313533044270964}
   - component: {fileID: 8437023491213692339}
   m_Layer: 0
@@ -130,22 +129,6 @@ Rigidbody2D:
   m_SleepingMode: 0
   m_CollisionDetection: 1
   m_Constraints: 4
---- !u!70 &1932466399221208136
-CapsuleCollider2D:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 8437023491213692340}
-  m_Enabled: 1
-  m_Density: 1
-  m_Material: {fileID: 0}
-  m_IsTrigger: 0
-  m_UsedByEffector: 0
-  m_UsedByComposite: 0
-  m_Offset: {x: 0, y: 0}
-  m_Size: {x: 0.22, y: 0.45}
-  m_Direction: 0
 --- !u!61 &5934313533044270964
 BoxCollider2D:
   m_ObjectHideFlags: 0
@@ -186,6 +169,5 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   _moveSpeed: 10
   _jumpSpeed: 10
-  health: 100
+  _jumpTimes: 1
   _isGround: 0
-  _slider: {fileID: 0}
