--- Assets/Sciprts/PlayerControllerTEST.cs
@@ -1,123 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class PlayerControllerTEST : MonoBehaviour
-{
-    public float runSpeed;
-    public float jumpSpeed;
-    public float doulbJumpSpeed;
-    public float climbSpeed;
-
-    private Rigidbody2D myRigidbody;
-    private Animator myAnim;
-    private BoxCollider2D myFeet;
-    private bool isGround;
-    private bool canDoubleJump;
-
-    // Start is called before the first frame update
-    void Start()
-    {
-        myRigidbody = GetComponent<Rigidbody2D>();
-        myAnim = GetComponent<Animator>();
-        myFeet = GetComponent<BoxCollider2D>();
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-        Flip();
-        Run();
-        Jump();
-        CheckGround();
-        switchAnimation();
-    }
-
-    void CheckGround()
-    {
-        isGround = myFeet.IsTouchingLayers(LayerMask.GetMask("Ground"));
-        //Debug.Log(isGround);
-    }
-
-    void Flip()
-    {
-        bool playerHasXAxisSpeed = Mathf.Abs(myRigidbody.velocity.x) > Mathf.Epsilon;
-        if (playerHasXAxisSpeed)
-        {
-            if (myRigidbody.velocity.x > 0.1f)
-            {
-                transform.localRotation = Quaternion.Euler(0, 0, 0);
-            }//方向往右不用翻轉
-            if (myRigidbody.velocity.x < -0.1f)
-            {
-                transform.localRotation = Quaternion.Euler(0, 180, 0);
-            }//往左翻轉
-        }
-    }
-
-
-    void Run()
-    {
-        float moveDir = Input.GetAxis("Horizontal");//移動監測方向水平軸
-        Vector2 playerVel = new Vector2(moveDir * runSpeed, myRigidbody.velocity.y);
-        myRigidbody.velocity = playerVel;
-        bool playerHasXAxisSpeed = Mathf.Abs(myRigidbody.velocity.x) > Mathf.Epsilon;
-        myAnim.SetBool("Run", playerHasXAxisSpeed);
-    }
-
-    void Jump()
-    {
-        if (Input.GetButtonDown("Jump"))
-        {
-            if (isGround)
-            {
-                myAnim.SetBool("Jump", true);
-                Vector2 jumpVel = new Vector2(0.0f, jumpSpeed);
-                myRigidbody.velocity = Vector2.up * jumpVel;
-                canDoubleJump = true;
-            }
-            else
-            {
-                if (canDoubleJump)
-                {
-                    myAnim.SetBool("DoubleJump", true);
-                    Vector2 doubleJumpVel = new Vector2(0.0f, doulbJumpSpeed);
-                    myRigidbody.velocity = Vector2.up * doubleJumpVel;
-                    canDoubleJump = false;
-                }
-            }
-        }
-    }
-
-    void switchAnimation()
-    {
-        myAnim.SetBool("Idle", false);
-        if (myAnim.GetBool("Jump"))
-        {
-            if (myRigidbody.velocity.y < 0.0f)
-            {
-                myAnim.SetBool("Jump", false);
-                myAnim.SetBool("Fall", true);
-            }
-        }
-        else if (isGround)
-        {
-            myAnim.SetBool("Fall", false);
-            myAnim.SetBool("Idle", true);
-        }
-        if (myAnim.GetBool("DoubleJump"))
-        {
-            if (myRigidbody.velocity.y < 0.0f)
-            {
-                myAnim.SetBool("DoubleJump", false);
-                myAnim.SetBool("DoubleFall", true);
-            }
-        }
-        else if (isGround)
-        {
-            myAnim.SetBool("DoubleFall", false);
-            myAnim.SetBool("Idle", true);
-        }
-    }
-}
-
