--- Assets/Sciprts/PlayerControllerTEST.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4298197ae16241c42bf9309c8072b55a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
