+++ Assets/Sciprts/Test.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7c3b93f48950d0f43b3d1dfbfd37f80e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
