+++ Assets/Sciprts/Test/CameraTest.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 03862311cfc5c27439c12b76fbc63514
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
